Updated headers Moved get singularity path to singularity config
