+++ lithops/serverless/backends/singularity/config.py
@@ -1,5 +1,5 @@
 #
-# (C) Copyright Cloudlab URV 2020
+# (C) Copyright Cloudlab URV 2024
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -15,6 +15,7 @@
 #
 
 import os
+import shutil
 
 FH_ZIP_LOCATION = os.path.join(os.getcwd(), 'lithops_singularity.zip')
 
@@ -57,6 +58,13 @@ SINGULARITYFILE_DEFAULT = """
 %runscript
     python3 /lithops/lithopsentry.py $AMQP_URL
 """
+
+
+def get_singularity_path():
+    singularity_path = shutil.which('singularity')
+    if not singularity_path:
+        raise Exception('singularity command not found. Install singularity')
+    return singularity_path
 
 
 def load_config(config_data):
